refactored tests out of the bigtable folder
